+++ src/main/java/com/together/buytogether/enroll/feature/JoinBuying.java
@@ -1,0 +1,42 @@
+package com.together.buytogether.enroll.feature;
+
+import com.together.buytogether.enroll.domain.Enroll;
+import com.together.buytogether.enroll.domain.EnrollRepository;
+import com.together.buytogether.member.domain.Member;
+import com.together.buytogether.member.domain.MemberRepository;
+import com.together.buytogether.member.domain.SessionConst;
+import com.together.buytogether.post.domain.Post;
+import com.together.buytogether.post.domain.PostRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+public class JoinBuying {
+    PostRepository postRepository;
+    EnrollRepository enrollRepository;
+    MemberRepository memberRepository;
+
+    public JoinBuying(
+            PostRepository postRepository,
+            EnrollRepository enrollRepository,
+            MemberRepository memberRepository) {
+        this.postRepository = postRepository;
+        this.enrollRepository = enrollRepository;
+        this.memberRepository = memberRepository;
+    }
+
+    @PostMapping("/posts/{postId}/enroll")
+    @ResponseStatus(HttpStatus.CREATED)
+    public void requset(
+            @SessionAttribute(name = SessionConst.LOGIN_MEMBER) Long memberId,
+            @PathVariable Long postId) {
+        Post post = postRepository.getByPostId(postId);
+        Member member = memberRepository.getByMemberId(memberId);
+        enrollRepository.findByMemberId(memberId).ifPresent(enroll -> {
+            throw new IllegalArgumentException("이미 참여한 구매글입니다.");
+        });
+        Enroll enroll = new Enroll(member, post);
+        post.increaseJoinCount();
+        enrollRepository.save(enroll);
+    }
+}
